third_party/blink/web_tests/wpt_internal/sanitizer-api/sanitizer-config.tentative.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>

<body>
<script>
    const default_option = {};
    test(t => {
      let s = new Sanitizer();
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, default_option);
    }, "SanitizerAPI creator without config.");

    test(t => {
      let s = new Sanitizer({});
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, default_option);
    }, "SanitizerAPI creator with empty config.");

    test(t => {
      let s = new Sanitizer(null);
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, default_option);
    }, "SanitizerAPI creator with null as config.");

    test(t => {
      let s = new Sanitizer(undefined);
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, default_option);
    }, "SanitizerAPI creator with undefined as config.");

    test(t => {
      let s = new Sanitizer({testConfig: [1,2,3], attr: ["test", "i", "am"]});
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, default_option);
    }, "SanitizerAPI creator with config ignore unknown values.");

    const config_names = ["dropElements", "dropAttributes"];
    config_names.forEach(cname => {
      test(t => {
        let options = {};
        options[cname] = cname.endsWith("Elements") ? ["div"] : ["script"];
        let result = {};
        result[cname] = cname.endsWith("Elements") ? ["DIV"] : ["script"];

        let s = new Sanitizer(options);
        assert_true(s instanceof Sanitizer);
        assert_object_equals(s.creationOptions, result);

        options[cname].push("test");
        assert_object_equals(s.creationOptions, result);

        let new_options = {};
        new_options[cname] = ["test", "t"];
        s.creationOptions = new_options;
        assert_object_equals(s.creationOptions, result);

        s.creationOptions[cname] = [1,2,3];
        assert_object_equals(s.creationOptions, result);
      }, "SanitizerAPI config " + cname + " is not editable.");

      let options = {};
      options[cname] = [];
      test(t => {
        let s = new Sanitizer(options);
        assert_true(s instanceof Sanitizer)
        assert_object_equals(s.creationOptions, options);
        assert_equals(s.sanitizeToString("<div>balabala<i>test</i></div>"), "<div>balabala<i>test</i></div>");
      }, "SanitizerAPI creator with config " + JSON.stringify(options) + ".");

      options = {};
      options[cname] = undefined;
      test(t => {
        let s = new Sanitizer(options);
        assert_true(s instanceof Sanitizer)
        assert_object_equals(s.creationOptions, default_option);
        assert_equals(s.sanitizeToString("<div>balabala<i>test</i></div>"), "<div>balabala<i>test</i></div>");
      }, "SanitizerAPI creator with config " + JSON.stringify(options, function(k,v){return v===undefined?"::undefined::":v}).replace(new RegExp("\"::undefined::\"", 'g'), "undefined") + ".");

      let testcases = [null, 123, "div"];
      testcases.forEach(c => {
        options = {};
        options[cname] = c;
        test(t => {
          assert_throws_js(TypeError, _ => {let s = new Sanitizer(options)});
        }, "SanitizerAPI creator with config " + JSON.stringify(options) + ".");
      });
    });

</script>
</body>
</html>
